% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_map}
\alias{area_map}
\title{Area Maps}
\format{
## `area_map`
\describe{
  \item{region_name}{Region name}
  \item{area_name}{Area name}
  \item{geometry}{geometry}
  \item{cent_long}{Longitude of centroid}
  \item{cent_lat}{Latitude of centroid}
  ...
}
}
\source{
<https://www.vic.gov.au/regional-model-department-education>
}
\usage{
area_map
}
\description{
Simple feature collection with 4 features and 4 fields
}
\keyword{datasets}
