% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_tbl}
\alias{region_tbl}
\title{Region Table}
\format{
## `region_tbl`
\describe{
  \item{region_name}{Region name, with four values}
  \item{area_name}{Area name, with 17 values}
  \item{lga_name}{Local government area name, with 79 values}
  \item{cent_long}{Longitude of centroid}
  \item{cent_lat}{Latitude of centroid}
  ...
}
}
\source{
<https://www.vic.gov.au/regional-model-department-education>
}
\usage{
region_tbl
}
\description{
A tibble of with 3 columns and 79 rows.
}
\keyword{datasets}
